refactor: 룰렛 서비스 -> 북마크 서비스로 통합, 컨트롤러로 도메인 분할
